CrawlerYotube/requirements.txt
pyautogui==0.9.52
pyperclip==1.8.2
pywin32==306
PyDirectInput==1.0.4
psycopg2-binary==2.9.1
requests==2.26.0
beautifulsoup4==4.9.3
Pillow==8.3.2
opencv-python==4.7.0.72
telegram==0.01
psutil==5.9.5